feat(tree): add left menu and an option to control availability of menus
Add setting's option - rightMenu which allows to control the availability of
the right menu. By default it is true, which mean that right menu is available.
When it is set to false there is no right menu.

Add left menu and an option to control it. This menu is usefull when the tree
should be displayed under mobile and tablet devices which don't have right
click. By default left menu isn't available.

Change event listeners for closing menus from click to mouse down, because
opening of a left menu is an left mouse event (click event) and if you add
event listener for closing menu on a click (when a menu is initialized), it
fires right after opening, which cause immediate closing.

Add a verification which on a mousedown event a close action will proceed only
if the event target is out of the menu's items, because other wise the menu is
destroyed before an action attached to the item is called.
Add a call to a closing function at the end of the fired event on selecting a
menu's item, because when a click event it trigger on a menu's item it
doesn't call the close function.